Fixed a minor bug with the PropertiesFilePreferences implementation that would unnecessarily create directories corresponding to Preferences nodes that were created in memory, loading from disk was attempted, but flushing entries back to disk never occurred.
For example, retrieve the node "/foo/bar/baz" to see if any entries
exist. Say that the folder foo exists, but that bar does not (and as
such the file backing the baz node). If no entries are placed into the
preferences node for /foo/bar/baz, there will be no need to flush the
node to disk. The bar directory should not be created in this instance.